Split EqualityComparer files
