+++ src/core-taggeds-optional/Optional/Optional.T/Optional.T.EqualityComparer.cs
@@ -1,12 +1,11 @@
 ﻿using System.Collections.Generic;
-using System.Runtime.CompilerServices;
 
 namespace System;
 
 partial struct Optional<T>
 {
     // TODO: Add the tests and open the class
-    internal sealed class EqualityComparer : IEqualityComparer<Optional<T>>
+    internal sealed partial class EqualityComparer : IEqualityComparer<Optional<T>>
     {
         private readonly IEqualityComparer<T> comparer;
 
@@ -25,30 +24,6 @@ partial struct Optional<T>
         public static EqualityComparer Default
             =>
             InnerDefault.Value;
-
-        public bool Equals(Optional<T> x, Optional<T> y)
-        {
-            if (x.hasValue != y.hasValue)
-            {
-                return false;
-            }
-
-            if (x.hasValue)
-            {
-                return comparer.Equals(x.value, y.value);
-            }
-
-            return true;
-        }
-
-        public int GetHashCode(Optional<T> obj)
-            =>
-            obj.hasValue ? PresentHashCode(obj.value) : default;
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private int PresentHashCode(T value)
-            =>
-            value is not null ? HashCode.Combine(comparer.GetHashCode(value)) : 1;
 
         private static class InnerDefault
         {
