+++ src/core-taggeds-result/Result/Result.T/Result.T.EqualityComparer.cs
@@ -1,12 +1,11 @@
 ﻿using System.Collections.Generic;
-using System.Runtime.CompilerServices;
 
 namespace System;
 
 partial struct Result<TSuccess, TFailure>
 {
     // TODO: Add the tests and open the class
-    internal sealed class EqualityComparer : IEqualityComparer<Result<TSuccess, TFailure>>
+    internal sealed partial class EqualityComparer : IEqualityComparer<Result<TSuccess, TFailure>>
     {
         private readonly IEqualityComparer<TSuccess> successComparer;
 
@@ -49,34 +48,6 @@ partial struct Result<TSuccess, TFailure>
         public static EqualityComparer Default
             =>
             InnerDefault.Value;
-
-        public bool Equals(Result<TSuccess, TFailure> x, Result<TSuccess, TFailure> y)
-        {
-            if (x.isSuccess != y.isSuccess)
-            {
-                return false;
-            }
-
-            return x.isSuccess
-                ? successComparer.Equals(x.success, y.success)
-                : failureComparer.Equals(x.failure, y.failure);
-        }
-
-        public int GetHashCode(Result<TSuccess, TFailure> obj)
-            =>
-            obj.isSuccess ? SuccessHashCode(obj.success) : FailureHashCode(obj.failure);
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private int SuccessHashCode(TSuccess success)
-            =>
-            success is not null
-                ? HashCode.Combine(true, successComparer.GetHashCode(success))
-                : HashCode.Combine(true);
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private int FailureHashCode(TFailure failure)
-            =>
-            HashCode.Combine(false, failureComparer.GetHashCode(failure));
 
         private static class InnerDefault
         {
